fix(gpu_prover): add disable delegation caching option and enable it (#69)
## What ❔

This PR adds a disable delegation caching option and enables it.

## Why ❔

There is an underlying race condition causing trace corruption when
delegation caching is enabled and multiple GPUs are used but so far the
cause us unclear. For now disabling delegation trace caching seems to
suppress it.

## Is this a breaking change?
- [ ] Yes
- [x] No

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [x] Code has been formatted.