Make FlattenTwo CLI work with ProgramProof
Previously, you had to have data on disk in a specific format (path to
ProofMetadata was provided via CLI & ProofList was expected implicitly
in the same parent as ProofMetadata, under a specific name).

Given most of the work/integration will be done via ProgramProof, this
PR changes the API to work with ProgramProof.